% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/void_dev.R
\name{void_dev}
\alias{void_dev}
\title{Open a void device}
\usage{
void_dev()
}
\description{
This function opens a graphic device that accepts inputs, but do not perform
any action. As such it is useful for benchmarking graphic code as it removes
the device implementation from the implementation.
}
\examples{
void_dev()
# Any plotting instructions will now have no effect
plot(1:10000, 1:10000)
dev.off()
}
