% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voidev-package.R
\docType{package}
\name{voidev-package}
\alias{voidev}
\alias{voidev-package}
\title{voidev: A Graphic Device That Does Nothing}
\description{
In order to properly benchmark graphic drawing code it is necessary
    to factor out the device implemenation itself so that results are not 
    related to the specifc graphics device used during benchmarking. The 
    'voidev' package implements a graphic device that accepts all the required
    calls from R's graphic engine but performs no action. Apart from 
    benchmarking it is unlikely that this device has any practical use.
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
