% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoid-package.R
\docType{package}
\name{devoid-package}
\alias{devoid}
\alias{devoid-package}
\title{devoid: A Graphic Device that Does Nothing}
\description{
Provides a non-drawing graphic device for benchmarking purpose.
    In order to properly benchmark graphic drawing code it is necessary
    to factor out the device implementation itself so that results are not 
    related to the specific graphics device used during benchmarking. The 
    'devoid' package implements a graphic device that accepts all the required
    calls from R's graphic engine but performs no action. Apart from 
    benchmarking it is unlikely that this device has any practical use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/devoid}
  \item Report bugs at \url{https://github.com/r-lib/devoid/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
